requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
aiomysql>=0.2.0
argon2-cffi>=21.0.0
PyJWT>=2.8.0
jinja2>=3.1.0
python-multipart>=0.0.5
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
aiosmtplib>=2.0.0
